third_party/blink/web_tests/editing/inserting/insert-3786362-fix-expected.txt

EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
This demonstrates a bug: copying an empty line that's implemented with a br doesn't put an interchange newline on the pasteboard.  There should be two blank lines after the 'a'.

Before paste:
| "
"
| <div>
|   class="editing"
|   "a"
|   <br>
|   <br>
|   "b"
| "
"
| <div>
|   class="editing"
|   <#selection-caret>
| "
"

After paste:
| "
"
| <div>
|   class="editing"
|   "a"
|   <br>
|   <br>
|   "b"
| "
"
| <div>
|   class="editing"
|   "a"
|   <br>
|   <#selection-caret>
|   <br>
| "
"
